Moved setup from server to root file
