+++ src/server.ts
@@ -1,25 +1,4 @@
-// read env vars from local .env file
-import * as dotenv from 'dotenv';
-dotenv.config();
-
-// check for required env vars declared in app.json
-import * as fs from 'fs';
-const appJSON = fs.readFileSync('app.json', 'utf8');
-const appConf = JSON.parse(appJSON);
-
-Object.keys(appConf.env).forEach((key) => {
-  let unset = false;
-  if (appConf.env[key].required && process.env[key] === undefined) {
-    console.error(key, 'is undefined!');
-    unset = true;
-  }
-  if (unset) {
-    console.error('You should set all env vars!');
-    console.error('For details check app.json and README');
-    process.exit(1);
-  }
-});
-
+import '../setup';
 import './controllers';
 import * as express from 'express';
 import * as path from 'path';
